# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "summary": "Create a customer account and sign in.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "summary": "Sign in with email + password.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "summary": "Return the authenticated user.",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/admin/categories": {
      "get": {
        "operationId": "CategoriesController_adminTree",
        "summary": "Admin category tree (same shape).",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "categories"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "CategoriesController_create",
        "summary": "Create a category.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "categories"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/categories/{id}": {
      "put": {
        "operationId": "CategoriesController_update",
        "summary": "Update a category.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "categories"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "CategoriesController_remove",
        "summary": "Delete a category (children re-parent to root).",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "categories"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/coupons/validate": {
      "post": {
        "operationId": "CouponsController_validate",
        "summary": "Validate a coupon code against a subtotal (cents).",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateCouponDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "coupons"
        ]
      }
    },
    "/api/admin/coupons": {
      "get": {
        "operationId": "CouponsController_list",
        "summary": "List coupons.",
        "parameters": [
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Field to sort by (allow-listed per resource).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "schema": {
              "default": "desc",
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "coupons"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "CouponsController_create",
        "summary": "Create a coupon.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCouponDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "coupons"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/coupons/{id}": {
      "put": {
        "operationId": "CouponsController_update",
        "summary": "Update a coupon.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCouponDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "coupons"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "CouponsController_remove",
        "summary": "Delete a coupon.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "coupons"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/cart/items": {
      "post": {
        "operationId": "CartController_add",
        "summary": "Add a variant to the cart.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "cart"
        ]
      }
    },
    "/api/cart/items/{itemId}": {
      "patch": {
        "operationId": "CartController_update",
        "summary": "Update a line item quantity (0 removes it).",
        "parameters": [
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cart"
        ]
      },
      "delete": {
        "operationId": "CartController_remove",
        "summary": "Remove a line item.",
        "parameters": [
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cart"
        ]
      }
    },
    "/api/cart/coupon": {
      "post": {
        "operationId": "CartController_applyCoupon",
        "summary": "Apply a coupon code to the cart.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyCouponDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "cart"
        ]
      },
      "delete": {
        "operationId": "CartController_removeCoupon",
        "summary": "Remove the applied coupon.",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cart"
        ]
      }
    },
...
  },
  "info": {
    "title": "Nexlor Commerce API",
    "description": "Storefront + Admin API for the Nexlor commerce template.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "jane@example.com"
          },
          "name": {
            "type": "string",
            "example": "Jane Doe"
          },
          "password": {
            "type": "string",
            "example": "Password123!",
            "minLength": 8
          }
        },
        "required": [
          "email",
          "name",
          "password"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "customer@nexlor.test"
          },
          "password": {
            "type": "string",
            "example": "Password123!"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string"
          },
          "parentId": {
            "type": "object",
            "description": "Parent category id for nesting."
          },
          "position": {
            "type": "number"
          }
        }
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string"
          },
          "parentId": {
            "type": "object",
            "description": "Parent category id for nesting."
          },
          "position": {
            "type": "number"
          }
        }
      },
...
      "ValidateCouponDto": {
        "type": "object",
        "properties": {}
      },
      "CreateCouponDto": {
        "type": "object",
        "properties": {
          "value": {
            "type": "number",
            "description": "PERCENT: 0–100. FIXED: cents."
          },
          "active": {
            "type": "boolean",
            "default": true
          },
          "expiresAt": {
            "type": "object",
            "description": "ISO date string."
          },
          "minSubtotal": {
            "type": "object",
            "description": "Minimum subtotal in cents."
          },
          "usageLimit": {
            "type": "object"
          }
        }
      },
      "UpdateCouponDto": {
        "type": "object",
        "properties": {
          "value": {
            "type": "number",
            "description": "PERCENT: 0–100. FIXED: cents."
          },
          "active": {
            "type": "boolean",
            "default": true
          },
          "expiresAt": {
            "type": "object",
            "description": "ISO date string."
          },
          "minSubtotal": {
            "type": "object",
            "description": "Minimum subtotal in cents."
          },
          "usageLimit": {
            "type": "object"
          }
        }
      },
      "AddItemDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateItemDto": {
        "type": "object",
        "properties": {}
      },
      "ApplyCouponDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}